+++ packages/entity-storage-connector-scylladb/tests/scyllaDBConnector.spec.ts
@@ -3,7 +3,7 @@
 /* eslint-disable max-classes-per-file */
 /* eslint-disable jsdoc/require-jsdoc */
 
-import { I18n } from "@gtsc/core";
+import { I18n } from "@twin.org/core";
 import {
 	ComparisonOperator,
 	EntitySchemaFactory,
@@ -11,16 +11,16 @@ import {
 	SortDirection,
 	entity,
 	property
-} from "@gtsc/entity";
-import { MemoryEntityStorageConnector } from "@gtsc/entity-storage-connector-memory";
-import { EntityStorageConnectorFactory } from "@gtsc/entity-storage-models";
+} from "@twin.org/entity";
+import { MemoryEntityStorageConnector } from "@twin.org/entity-storage-connector-memory";
+import { EntityStorageConnectorFactory } from "@twin.org/entity-storage-models";
 import {
 	EntityStorageLoggingConnector,
 	type LogEntry,
 	initSchema
-} from "@gtsc/logging-connector-entity-storage";
-import { LoggingConnectorFactory } from "@gtsc/logging-models";
-import { nameof } from "@gtsc/nameof";
+} from "@twin.org/logging-connector-entity-storage";
+import { LoggingConnectorFactory } from "@twin.org/logging-models";
+import { nameof } from "@twin.org/nameof";
 import type { IScyllaDBTableConfig } from "../src/models/IScyllaDBTableConfig";
 import { ScyllaDBTableConnector } from "../src/scyllaDBTableConnector";