+++ packages/entity-storage-connector-file/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/entity-storage-connector-file - Changelog
+# @twin.org/entity-storage-connector-file - Changelog
 
-## v0.0.3
+## v0.0.1-next.2
 
 - Initial Release