+++ packages/entity-storage-connector-scylladb/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/entity-storage-connector-scylladb - Changelog
+# @twin.org/entity-storage-connector-scylladb - Changelog
 
-## v0.0.3
+## v0.0.1-next.2
 
 - Initial Release